Fix panic in `impl Sum for FpVar`
`AllocatedFp::add_many` must not be called with an empty iterator.

CC #151

This fixes a bug for our project depending on the `ark-*` libraries that
surfaced when trying to upgrade to 0.5.0.

I also looked at whether one could make `AllocatedFp::add_many` not
panic in the empty case, but that wasn't as straightforward as many
operations on `AllocatedFp` depend on there being a constraint system
reference.